# backend/src/lib/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FinAnalysis API",
    "version": "1.0.0",
    "description": "FinAnalysis API - Financial analysis and advisory platform. All endpoints are prefixed with `/api/v1`. Uses RFC 7807 Problem Details for errors."
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "Current API version"
    },
    {
      "url": "/api",
      "description": "Legacy API (deprecated)"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token"
      }
    },
    "schemas": {
      "ProblemDetails": {
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "instance"
        ],
        "properties": {
          "type": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/ProblemDetails"
          }
        ]
      },
...
    },
    "responses": {
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "validationError": {
                "summary": "Validation error",
                "value": {
                  "type": "https://tools.ietf.org/html/rfc7807#section-3.1",
                  "title": "Validation Error",
                  "status": 400,
                  "detail": "email: Invalid email format",
                  "instance": "/auth/login"
                }
              },
              "unauthorized": {
                "summary": "Unauthorized",
                "value": {
                  "type": "https://tools.ietf.org/html/rfc7807#section-3.1",
                  "title": "Unauthorized",
                  "status": 401,
                  "detail": "Invalid or expired token",
                  "instance": "/portfolio"
                }
              },
              "notFound": {
                "summary": "Not found",
                "value": {
                  "type": "https://tools.ietf.org/html/rfc7807#section-3.1",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "Portfolio not found",
                  "instance": "/portfolio/550e8400-e29b-41d4-a716-446655440002"
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized - Invalid or missing authentication",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "validationError": {
                "summary": "Validation error",
                "value": {
                  "type": "https://tools.ietf.org/html/rfc7807#section-3.1",
                  "title": "Validation Error",
                  "status": 400,
                  "detail": "email: Invalid email format",
                  "instance": "/auth/login"
                }
              },
              "unauthorized": {
                "summary": "Unauthorized",
                "value": {
                  "type": "https://tools.ietf.org/html/rfc7807#section-3.1",
                  "title": "Unauthorized",
                  "status": 401,
                  "detail": "Invalid or expired token",
                  "instance": "/portfolio"
                }
              },
              "notFound": {
                "summary": "Not found",
                "value": {
                  "type": "https://tools.ietf.org/html/rfc7807#section-3.1",
                  "title": "Not Found",
                  "status": 404,
                  "detail": "Portfolio not found",
                  "instance": "/portfolio/550e8400-e29b-41d4-a716-446655440002"
                }
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/auth/refresh": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Refresh access token",
        "operationId": "refreshToken",
        "responses": {
          "200": {
            "description": "Token refreshed"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        },
        "security": [
          {
            "cookieAuth": []
          }
        ]
      }
    },
...
    "/auth/verify-otp": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Verify OTP",
        "operationId": "verifyOtp",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "otp"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "otp": {
                    "type": "string"
                  }
                }
              },
              "examples": {
                "valid": {
                  "summary": "Verify OTP",
                  "value": {
                    "email": "john.doe@example.com",
                    "otp": "123456"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OTP verified"
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
...
  }
}